split out vault page into compoennts
