+++ frontend/src/components/VaultsPage/index.ts
@@ -1,0 +1,4 @@
+export * from "./VaultPage";
+export * from "./VaultMetaDisplay";
+export * from "./VaultDeposit";
+export * from "./VaultWithdraw";
