+++ frontend/src/components/index.ts
@@ -1,4 +1,4 @@
 export * from "./Header";
 export * from "./VaultsTable";
 export * from "./common";
-export * from "./VaultPage";
+export * from "./VaultsPage/";
